+++ CHANGELOG.md
@@ -16,4 +16,4 @@
 ## 0.2.0
 
 * Fix some bugs on Connection handover records
-* Utilize extension methods to simplify usage
+* Use extension methods to simplify usage